src/requirements.txt
Django==4.2.6
asgiref==3.7.2
autopep8==1.6.0
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
django-environ==0.8.1
django-allauth==0.51.0
django-sslserver==0.22
dj-stripe==2.7.3
django-payments==2.0.0
openai==0.27.8
six==1.16.0
mysqlclient==2.1.1
uvicorn==0.22.0
httpx
